Move ProgressBar to components/ui and add storybook
