--- airbyte-webapp/src/components/ProgressBar/index.tsx
@@ -1,4 +1,0 @@
-import ProgressBar from "./ProgressBar";
-
-export default ProgressBar;
-export { ProgressBar };
