+++ airbyte-webapp/src/components/index.tsx
@@ -17,5 +17,4 @@ export * from "./PageTitle";
 export * from "./SimpleTableComponents";
 export * from "./StatusIcon";
 export * from "./TextWithHTML";
-export * from "./ProgressBar";
 export * from "./ConnectorCard";
