+++ airbyte-webapp/src/components/ui/ProgressBar/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ProgressBar";
